services/books_service/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
Deprecated==1.2.18
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flake8==7.2.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-tools==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
itsdangerous==2.2.0
jalali_core==1.0.0
jdatetime==5.2.0
Jinja2==3.1.6
limits==5.2.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mssql-django==1.5
mypy_extensions==1.1.0
nodeenv==1.9.1
orjson==3.10.18
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pre_commit==4.2.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
uvicorn==0.34.3
virtualenv==20.31.2
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
